feat: enhance Typst processing and update post template with metadata handling
