+++ posts/content/2026-03-10-hello-world.typ
@@ -7,6 +7,8 @@
 
 这是一篇测试文章，用于验证 Typst 的基本功能。
 
+こんにちは、世界！
+
 ```javascript
 console.log("Hello world");
 ```
